+++ docs/versionhistory.rst
@@ -14,6 +14,7 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
   or ``IPv6Interface`` if the address contains host bits
 - **BACKWARD INCOMPATIBLE** Changed the encoding of IP addresses to use the semantic tags 52 and 54
   instead of the deprecated 260 and 261 (`#232 <https://github.com/agronholm/cbor2/issues/232>`_)
+- Added configurable maximum depth for container nesting (via the ``max_depth`` parameter)
 - Fixed string references to work with indefinite-length (byte/unicode) strings too
 
 **UNRELEASED**